Add files via upload
To check for the session storage trimmed Email value and create a replacedEmail variable for the URL of the user dashboard. Also restricting the user from wandering around to the other users dashboards and redirects them to the 404 page if they try to do so. The user can only navigate to only their own dashboard and to the allowed URLs.